+++ docs/source/releasenotes.rst
@@ -6,7 +6,7 @@ Upcoming release
 
 - Library **Desktop**: Handle locators with whitespace, allow using return values as arguments
 - Library **Excel.Files**: Expose ``header`` argument in ``Create Worksheet``
-
+- Library **FileSystem**: Add keyword for reading file owner
 
 7.0.5
 -----